ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
openai==1.35.0
azure-cognitiveservices-speech==1.38.0
redis==5.0.7
pydantic==2.8.0
sentence-transformers==3.0.0
nltk==3.8.1
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.0
black==24.4.0
flake8==7.1.0